refactor(timeline): extract critical path logic into criticalPath.ts helper
