+++ packages/jaeger-ui/src/components/TracePage/TraceTimelineViewer/VirtualizedTraceView.test.jsx
@@ -513,60 +513,6 @@ describe('<VirtualizedTraceViewImpl>', () => {
       const localInstance = document.querySelector('.VirtualizedTraceView--spans');
       expect(localInstance).toBeInTheDocument();
     });
-
-    it('returns [] from mergeChildrenCriticalPath when criticalPath is falsy', () => {
-      const span = trace.spans[0];
-      const result = VirtualizedTraceViewImpl.prototype.getCriticalPathSections.call(
-        { props: { ...mockProps, trace } },
-        true,
-        false,
-        trace,
-        span,
-        undefined
-      );
-      expect(result).toEqual([]);
-    });
-
-    it('merges only pruned subtree critical path when hasPrunedChildren is true', () => {
-      const prunedChild = {
-        spanID: 'pruned-child',
-        hasChildren: false,
-        resource: { serviceName: 'svc-pruned' },
-        childSpans: [],
-      };
-      const visibleChild = {
-        spanID: 'visible-child',
-        hasChildren: false,
-        resource: { serviceName: 'svc-visible' },
-        childSpans: [],
-      };
-      const parentSpan = {
-        ...trace.spans[0],
-        spanID: 'parent',
-        hasChildren: true,
-        childSpans: [prunedChild, visibleChild],
-      };
-      // Create a trace whose spans array includes the parent so the memoized
-      // precomputation can find it.
-      const customTrace = { ...trace, spans: [parentSpan, prunedChild, visibleChild] };
-      const fakeCriticalPath = [
-        { spanID: 'parent', sectionStart: 0, sectionEnd: 10 },
-        { spanID: 'pruned-child', sectionStart: 10, sectionEnd: 20 },
-        { spanID: 'visible-child', sectionStart: 20, sectionEnd: 30 },
-      ];
-      const result = VirtualizedTraceViewImpl.prototype.getCriticalPathSections.call(
-        { props: { ...mockProps, trace: customTrace, prunedServices: new Set(['svc-pruned']) } },
-        false,
-        true,
-        customTrace,
-        parentSpan,
-        fakeCriticalPath
-      );
-      const spanIDs = result.map(s => s.spanID);
-      expect(spanIDs).toContain('parent');
-      expect(spanIDs).toContain('pruned-child');
-      expect(spanIDs).not.toContain('visible-child');
-    });
   });
 
   describe('shouldScrollToFirstUiFindMatch', () => {
